+++ test/server.js
@@ -10,6 +10,7 @@ let app = express(),
 	getTemplate = pug.compileFile(path.join(__dirname, './index.pug')),
 	template = getTemplate({bundleLink: `http://127.0.0.1:${config.devserverPort}/dist/main.js`})
 app.get('/', (req, res) => res.send(template))
+app.get('/static/:fileName', (req, res) => res.sendFile(path.join(__dirname, '/static', decodeURIComponent(req.params.fileName))))
 app.get('/testModel/selectList', (req, res) => res.json([
 	{text: 'Option 1', value: 1},
 	{text: 'Option 2', value: 2},